api display order history and infor user

+++ Models/Orders.cs
@@ -6,7 +6,7 @@
         public int id_customer { get; set; }
         public DateTime order_date { get; set; }
         public string shippingAddress { get; set; }
-        public int total_amount { get; set; }
+        public Decimal total_amount { get; set; }
         public string paymentMethod { get; set; }
         public int orderStatus { get; set; }
         public List<OrderItem> orderItems { get; set; }
